updated routes and controllers folder
-Created deleteIngredient function
