+++ backend/src/controllers/batchController.js
@@ -1,0 +1,14 @@
+/*
+  TODO //?changeCategoryName
+  TODO 
+  TODO addBatch
+  TODO deleteBatch (low priority for now)
+  TODO other batch methods (low priority for now)
+  TODO toggleCategoryExcluded
+  TODO toggleIngredientExcluded
+  TODO getEarliestExpiringBatch
+  TODO Add ingredient, category should always be specified, Uncategorized is a category
+  TODO Add expirationDate to a batch, 
+  TODO Implement batch merging logic
+       ?Should user be able to have the same ingredient in different categories?
+  */
